Refacto: Changed symbols on glyphs system to use ASCII-256 Char so we cant avoid using provided functions

+++ src/utils/misc_utils.c
@@ -6,7 +6,7 @@
 /*   By: vdurand <vdurand@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/21 18:07:59 by vdurand           #+#    #+#             */
-/*   Updated: 2025/09/19 17:22:49 by vdurand          ###   ########.fr       */
+/*   Updated: 2025/09/22 16:41:40 by vdurand          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,15 +35,15 @@ uint64_t	get_seed(void)
 	file_fd = open(SEED_SYSTEM_RAND_FILE, O_RDONLY);
 	if (file_fd == -1)
 	{
-		printf(WARNING_SEED, SEED_FALLBACK_DEFAULT);
-		return (SEED_FALLBACK_DEFAULT);
+		printf(WARNING_SEED, (uint64_t)SEED_FALLBACK_DEFAULT);
+		return ((uint64_t)SEED_FALLBACK_DEFAULT);
 	}
 	len = read(file_fd, &result, sizeof(result));
 	close(file_fd);
 	if (len < sizeof(result))
 	{
-		printf(WARNING_SEED, SEED_FALLBACK_DEFAULT);
-		return (SEED_FALLBACK_DEFAULT);
+		printf(WARNING_SEED, (uint64_t)SEED_FALLBACK_DEFAULT);
+		return ((uint64_t)SEED_FALLBACK_DEFAULT);
 	}
 	printf(SEED_MESSAGE, result);
 	return (result);
